Reorganized the router/handler setup to better reflect separation of concerns and modularity.
