+++ routers/handlers/healthz.go
@@ -1,0 +1,8 @@
+package handlers
+
+import "net/http"
+
+func Healthz(w http.ResponseWriter, r *http.Request) {
+	w.WriteHeader(http.StatusOK)
+	w.Write([]byte("OK"))
+}
