+++ routers/handlers/register_service_worker.go
