+++ routers/routers.go
@@ -5,8 +5,11 @@ import (
 
 	"github.com/go-chi/chi/v5"
 	"github.com/regcomp/gdpr/config"
-	"github.com/regcomp/gdpr/handlers"
 	"github.com/regcomp/gdpr/middleware"
+	"github.com/regcomp/gdpr/routers/api"
+	authproxy "github.com/regcomp/gdpr/routers/auth_proxy"
+	"github.com/regcomp/gdpr/routers/client"
+	"github.com/regcomp/gdpr/routers/handlers"
 	servicecontext "github.com/regcomp/gdpr/service_context"
 )
 
@@ -22,7 +25,7 @@ func CreateRouter(stx *servicecontext.ServiceContext) *chi.Mux {
 		middleware.TraceRequests,
 	)
 
-	router.Get(config.EndpointHealthz, healthz)
+	router.Get(config.EndpointHealthz, handlers.Healthz)
 	router.Get(
 		config.EndpointRegisterServiceWorker,
 		handlers.RegisterServiceWorker(stx.RequestStore),
@@ -61,61 +64,26 @@ func CreateServiceRouter(stx *servicecontext.ServiceContext) SubRouter {
 	)
 
 	mountRouters(service,
-		CreateAuthRouter(stx),
-		CreateClientRouter(stx),
-		CreateAPIRouter(stx),
+		CreateAuthProxyRouter(stx),
+		CreateRequiresValidAuthRouter(stx),
 	)
 	return SubRouter{"/", service}
 }
 
-func CreateAuthRouter(stx *servicecontext.ServiceContext) SubRouter {
-	auth := chi.NewRouter()
+func CreateRequiresValidAuthRouter(stx *servicecontext.ServiceContext) SubRouter {
+	requiresValidAuth := chi.NewRouter()
 
-	auth.Use(
-		middleware.SkipIfAuthenticated(stx.AuthProvider, stx.CookieManager, stx.ConfigStore),
-	)
-
-	auth.Get(config.EndpointLogin, handlers.LoginPage)
-	auth.Post(config.EndpointLogin, handlers.SubmitLoginCredentials(stx.AuthProvider, stx.ConfigStore))
-
-	// Apparently some providers will hit with either GET or POST
-	loginCallback := handlers.LoginCallback(stx.AuthProvider, stx.CookieManager, stx.SessionStore)
-	auth.Get(config.EndpointLoginCallback, loginCallback)
-	auth.Post(config.EndpointLoginCallback, loginCallback)
-
-	auth.Post(config.EndpointRenewToken, handlers.RenewAccessToken(stx.AuthProvider, stx.CookieManager))
-
-	return SubRouter{Path: config.RouterAuthPathPrefix, Router: auth}
-}
-
-func CreateClientRouter(stx *servicecontext.ServiceContext) SubRouter {
-	client := chi.NewRouter()
-
-	client.Use(
-		middleware.RequiresAuthentication(stx.AuthProvider, stx.CookieManager),
-		middleware.HasActiveSession(stx.SessionStore, stx.CookieManager),
-		middleware.AddNonceToRequest(stx.NonceStore),
-	)
-
-	client.Get(config.EndpointDashboard, handlers.DashboardPage(stx.CookieManager))
-	client.Get(config.EndpointTest1, handlers.TestEndpoint1)
-	client.Get(config.EndpointTest2, handlers.TestEndpoint2)
-
-	// NOTE: Not sure were this should go
-	client.Post(config.EndpointLogout, handlers.Logout(stx.CookieManager))
-
-	return SubRouter{Path: config.RouterClientPathPrefix, Router: client}
-}
-
-func CreateAPIRouter(stx *servicecontext.ServiceContext) SubRouter {
-	api := chi.NewRouter()
-
-	api.Use(
+	requiresValidAuth.Use(
 		middleware.RequiresAuthentication(stx.AuthProvider, stx.CookieManager),
 		middleware.HasActiveSession(stx.SessionStore, stx.CookieManager),
 	)
 
-	return SubRouter{Path: config.RouterApiPathPrefix, Router: api}
+	mountRouters(requiresValidAuth,
+		CreateClientRouter(stx),
+		CreateAPIRouter(stx),
+	)
+
+	return SubRouter{Path: "/", Router: requiresValidAuth}
 }
 
 func mountRouters(main *chi.Mux, subrouters ...SubRouter) {
@@ -128,7 +96,23 @@ func mountRouters(main *chi.Mux, subrouters ...SubRouter) {
 	}
 }
 
-func healthz(w http.ResponseWriter, r *http.Request) {
-	w.WriteHeader(http.StatusOK)
-	w.Write([]byte("OK"))
+func CreateAuthProxyRouter(stx *servicecontext.ServiceContext) SubRouter {
+	return SubRouter{
+		Path:   config.RouterAuthPathPrefix,
+		Router: authproxy.CreateAuthProxyRouter(stx),
+	}
+}
+
+func CreateClientRouter(stx *servicecontext.ServiceContext) SubRouter {
+	return SubRouter{
+		Path:   config.RouterClientPathPrefix,
+		Router: client.CreateClientRouter(stx),
+	}
+}
+
+func CreateAPIRouter(stx *servicecontext.ServiceContext) SubRouter {
+	return SubRouter{
+		Path:   config.RouterApiPathPrefix,
+		Router: api.CreateApiRouter(stx),
+	}
 }
